platform-implement-win.cpp: Fix invalid pointer cast in 64 bits systems
The ID returned by the function SetTimer is actually a pointer to a data
structure storing the timer information. Originally the return type of
the function was int, which is 32 bits, but in Windows 64 bits it is not
enough to store a pointer value.

We convert the timer ID type to the proper type intptr_t, which is an
integral type guaranteed to be big enough to hold a pointer value.

+++ dali/internal/window-system/windows/platform-implement-win.h
@@ -45,9 +45,9 @@ bool PostWinThreadMessage(
 
 using timerCallback = bool(*)(void *data);
 
-int SetTimer(int interval, timerCallback callback, void *data);
+intptr_t SetTimer(int interval, timerCallback callback, void *data);
 
-void KillTimer(int id);
+void KillTimer(intptr_t id);
 
 const char* GetKeyName( int keyCode );
